# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Account:
    properties:
      account_currency:
        $ref: '#/definitions/domain.Currency'
      account_currency_id:
        type: string
      account_number:
        type: string
      active:
        type: boolean
      balance:
        type: number
      created_at:
        type: string
      entity:
        $ref: '#/definitions/domain.Entity'
      entity_id:
        type: string
      id:
        type: string
      opened_at:
        type: string
      type:
        $ref: '#/definitions/domain.AccountType'
      updated_at:
        type: string
    required:
    - account_number
    type: object
  domain.AccountType:
    enum:
    - TRADING
    - SETTLEMENT
    - CUSTODY
    - MARGIN
    type: string
    x-enum-varnames:
    - AccountTypeTrading
    - AccountTypeSettlement
    - AccountTypeCustody
    - AccountTypeMargin
...
  handler.succeedProvisionalLEIRequest:
    properties:
      official_lei:
        type: string
    required:
    - official_lei
    type: object
  handler.translationListResponse:
    properties:
      records:
        items:
          $ref: '#/definitions/handler.uiTranslationPublicDTO'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@axiom.example.com
    name: API Support
  description: Financial Services Static Data Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Axiom API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Get list of all accounts
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Account'
            type: array
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new account
      parameters:
      - description: Account object
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/domain.Account'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Get a single account by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Account'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account by ID
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update an existing account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Account object
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/domain.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update account
      tags:
      - accounts
  /admin/translations:
    get:
      description: Return a paginated list of translation strings, optionally filtered
        by language, status, or search text
      parameters:
      - description: Filter by ISO 639-1 language code (e.g. fr, es)
        in: query
        name: language
        type: string
      - description: Filter by status (pending, approved, rejected)
        in: query
        name: status
        type: string
      - description: Search by key or value substring
        in: query
        name: search
        type: string
      - description: Maximum records to return (default 50)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.translationListResponse'
      security:
      - BearerAuth: []
      summary: List UI translations for admin review
      tags:
      - translations
  /api/v1/lei:
    get:
      consumes:
      - application/json
      description: Get a paginated list of LEI records with optional search and filters
      parameters:
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - description: Search term (LEI code or legal name)
        in: query
        name: search
        type: string
      - description: Entity status filter (e.g., ACTIVE, INACTIVE)
        in: query
        name: status
        type: string
      - description: Entity category filter (e.g., GENERAL, FUND)
        in: query
        name: category
        type: string
      - description: Country code filter (e.g., US, GB)
        in: query
        name: country
        type: string
      - description: Sort field (lei, legal_name, entity_status, entity_category,
          legal_address_country, last_update_date)
        in: query
        name: sortBy
        type: string
      - default: asc
        description: Sort order (asc, desc)
        in: query
        name: sortOrder
        type: string
      - description: Include successor/managing LOU legal names in list response
        in: query
        name: includeLinkedNames
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LEIRecord'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List LEI records
      tags:
      - LEI
  /api/v1/lei-categories:
    get:
      description: Get sorted list of unique category values from LEI records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get distinct categories
      tags:
      - LEI
  /api/v1/lei-countries:
    get:
      description: Get sorted list of unique countries from LEI records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get distinct countries
      tags:
      - LEI
  /api/v1/lei-legal-forms:
    get:
      description: Get sorted list of unique legal form values from LEI records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get distinct legal forms
      tags:
      - LEI
  /api/v1/lei-regions:
    get:
      description: Get sorted list of unique regions from LEI records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get distinct regions
      tags:
      - LEI
  /api/v1/lei/{lei}:
    get:
      consumes:
      - application/json
      description: Get a specific LEI record by its LEI code
      parameters:
      - description: LEI code
        in: path
        name: lei
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.LEIRecord'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get LEI record by code
      tags:
      - LEI
  /api/v1/lei/{lei}/audit:
    get:
      consumes:
      - application/json
      description: Get audit trail for a specific LEI record
      parameters:
      - description: LEI code
        in: path
        name: lei
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LEIRecordAudit'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get LEI audit history
      tags:
      - LEI
  /api/v1/lei/{lei}/predecessors:
    get:
      consumes:
      - application/json
      description: Get LEI records whose successor_lei equals the provided LEI
      parameters:
      - description: LEI code
        in: path
        name: lei
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LEIRecord'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get predecessor LEI records
      tags:
      - LEI
  /api/v1/lei/count:
    get:
      description: Returns the total count of LEI records stored in the database,
        regardless of sync status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: int64
              type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get total LEI record count
      tags:
      - LEI
  /api/v1/lei/import-failures:
    get:
      consumes:
      - application/json
      description: Get persisted Level 1/Level 2 processing failures with open/resolved
        lifecycle filtering
      parameters:
      - description: Job type filter (DAILY_FULL, DAILY_DELTA, LEVEL2_RR, LEVEL2_REPEX)
        in: query
        name: jobType
        type: string
      - default: true
        description: When true (default), returns only unresolved failures
        in: query
        name: openOnly
        type: boolean
      - default: 100
        description: Max rows
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List LEI import processing failures
      tags:
      - LEI
  /api/v1/lei/level2/failures:
    get:
      consumes:
      - application/json
      description: Deprecated endpoint. Use /api/v1/lei/import-failures instead.
      parameters:
      - description: Job type filter (DAILY_FULL, DAILY_DELTA, LEVEL2_RR, LEVEL2_REPEX)
        in: query
        name: jobType
        type: string
      - default: true
        description: When true (default), returns only unresolved failures
        in: query
        name: openOnly
        type: boolean
      - default: 100
        description: Max rows
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List LEI import processing failures (Deprecated)
      tags:
      - LEI
  /api/v1/lei/names:
    get:
      description: Returns a map of LEI code to legal name. Input codes are trimmed,
        normalized, deduplicated, and validated (20-char uppercase alphanumeric).
        Codes not found are absent from the response.
      parameters:
      - description: Comma-separated list of LEI codes (max 500)
        in: query
        name: codes
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Batch-fetch legal names for a set of LEI codes
      tags:
      - LEI
  /api/v1/lei/provisional:
    get:
      description: Returns provisional LEI records with pagination.
      parameters:
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List provisional LEI records
      tags:
      - Provisional LEI
    post:
      consumes:
      - application/json
      description: Issues a new provisional LEI record.
      parameters:
      - description: Create provisional LEI request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateProvisionalLEIRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.LEIRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create provisional LEI record
      tags:
      - Provisional LEI
...
  /api/v1/lei/provisional/{lei}/succeed:
    post:
      consumes:
      - application/json
      description: Marks a provisional LEI as succeeded by a successor LEI.
      parameters:
      - description: Provisional LEI code
        in: path
        name: lei
        required: true
        type: string
      - description: Succeed provisional LEI request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.succeedProvisionalLEIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Succeed provisional LEI
      tags:
      - Provisional LEI
...
  /auth/users/{id}/approve:
    post:
      description: Activate a user account so the user can log in
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve a pending user (admin only)
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"